third_party/blink/web_tests/editing/shadow/shadow-selection-not-exported-expected.txt

CONSOLE WARNING: line 16: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
PASS assertNotInShadow(selection.anchorNode) is true
PASS assertNotInShadow(selection.focusNode) is true
PASS assertNotInShadow(selection.anchorNode) is true
PASS assertNotInShadow(selection.focusNode) is true
PASS assertNotInShadow(selection.anchorNode) is true
PASS assertNotInShadow(selection.focusNode) is true
PASS successfullyParsed is true

TEST COMPLETE


This test ensures shadow root won't expose to JS layer.

BEFORE
AFTER
